Refactored environment variables to use .env file
set docker-compose.yml env_file field on the web service. This points at a
.env file in the root directory to receive environment variables. The main
purpose use case is to allow for separate configurations in development and
production, and to keep secrets out of the git repository.

additionally this allows for the modified manage.py file to read these
variables so the application can be run outside of a docker container. The
primary use case for this, although it can be done with docker-compose, is
to run tests:

  python manage.py test --settings=api.config.settings